+++ src/states_screens/tracks_screen.cpp
@@ -141,7 +141,7 @@ void TracksScreen::beforeAddingWidget()
     tabs->clearAllChildren();
 
     const std::vector<std::string>& groups = track_manager->getAllTrackGroups();
-    const int group_amount = groups.size();
+    const int group_amount = (int)groups.size();
 
     if (group_amount > 1)
     {
@@ -161,7 +161,7 @@ void TracksScreen::beforeAddingWidget()
         tabs->addTextChild( _(groups[n].c_str()), groups[n] );
 
     DynamicRibbonWidget* tracks_widget = getWidget<DynamicRibbonWidget>("tracks");
-    tracks_widget->setItemCountHint( track_manager->getNumberOfTracks()+1 );
+    tracks_widget->setItemCountHint( (int)track_manager->getNumberOfTracks()+1 );
 }   // beforeAddingWidget
 
 // -----------------------------------------------------------------------------
@@ -256,7 +256,7 @@ void TracksScreen::buildTrackList()
 
     const std::string& curr_group_name = tabs->getSelectionIDString(0);
 
-    const int track_amount = track_manager->getNumberOfTracks();
+    const int track_amount = (int)track_manager->getNumberOfTracks();
 
     // First build a list of all tracks to be displayed
     // (e.g. exclude arenas, ...)
